Refactor #2 - Relocate function to parent Component
This would be much easier for maintainance in the future when the project gets into massive size.

If the functions tends to stay in child component, it's harsh to trackdown the child component.
