+++ src/components/ToDo.jsx
@@ -1,13 +1,21 @@
 import React, { useState } from 'react';
+import uuid from 'react-uuid';
 
 import Form from '../commons/Form';
 import List from './List';
 
 function ToDo() {
   const [list, setList] = useState([]);
+  const [input, setInput] = useState('');
 
-  function appendTask(value) {
-    setList([...list, value]);
+  function updateValue(event) {
+    setInput(event.target.value);
+  }
+
+  function handleSubmit(event) {
+    event.preventDefault();
+    setList([...list, { message: input, id: uuid() }]);
+    setInput('');
   }
 
   function detachTask(id) {
@@ -17,7 +25,7 @@ function ToDo() {
   return (
     <div>
       <h1>To-do</h1>
-      <Form appendTask={appendTask} />
+      <Form updateValue={updateValue} handleSubmit={handleSubmit} />
       <List list={list} onClick={detachTask} />
     </div>
   );
